:recycle: Delegated initialization to a dedicated object
