+++ lib/engines/core.rb
@@ -1,4 +1,5 @@
 # frozen_string_literal: true
+
 module Ducky
   module Engine
     class Core
@@ -16,13 +17,9 @@ module Ducky
         p '[Ducky] Core engine has been instantiated'
       end
 
-      def system_update(args)
-        @engines.each { |_key, engine| engine.system_update(args) }
-        @game.system_update(args)
-      end
-
-      def system_draw(args)
-        @game.system_draw(args)
+      def tick(args)
+        system_update(args)
+        system_draw(args)
       end
 
       def add_scene(scene_class)
@@ -30,9 +27,20 @@ module Ducky
       end
 
       def start_scene(scene, args)
-        raise MissingConfiguration, 'You must give at least on scene to Engine::Core' if scene.nil?
+        raise MissingConfiguration, 'You must give at least one scene to Engine::Core' if scene.nil?
 
         @current_scene = scene.new(args).tap(&:enter_tree)
+      end
+
+      private
+
+      def system_update(args)
+        @engines.each { |_key, engine| engine.system_update(args) }
+        @game.system_update(args)
+      end
+
+      def system_draw(args)
+        @game.system_draw(args)
       end
     end
   end
